\name{SixuanGo-package}
\alias{SixuanGo-package}
\alias{SixuanGo}
\docType{package}
\title{
\packageTitle{SixuanGo}
}
\description{
\packageDescription{SixuanGo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SixuanGo}
\packageIndices{SixuanGo}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{SixuanGo}

Maintainer: \packageMaintainer{SixuanGo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
